+++ CloudSchoolBusParent/etonkids-iphone/Class/notice/ETNoticeViewController.m
@@ -549,7 +549,7 @@
 
 #pragma mark --------- notice cell delegate -----------
 
-- (void) didTapImageWithImageArray:(NSArray *)imgArr showNumber:(int)num content:(NoticeInfo *)notice
+- (void) didTapImageWithImageArray:(NSArray *)imgArr showNumber:(NSInteger)num content:(NoticeInfo *)notice
 {
     if (imgArr != nil && notice != nil) {
         ETShowBigImageViewController *showBigVC = [[ETShowBigImageViewController alloc] init];
