+++ CloudSchoolBusParent/etonkids-iphone/Class/video/OpenGLView20.h
@@ -60,7 +60,7 @@
 #endif
 }
 #pragma mark - 接口
-- (void)displayYUV420pData:(void *)data width:(NSInteger)w height:(NSInteger)h;
+- (void)displayYUV420pData:(void *)data width:(int32_t)w height:(int32_t)h;
 - (void)setVideoSize:(GLuint)width height:(GLuint)height;
 
 /**
